client_side/pyproject.toml
[tool.poetry]
name = "client_side"
version = "0.1.0"
description = "simple python project to send transactions using the bridge contract"
authors = ["Taehono Kim <taehoon.kim@uzh.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^6.19.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
